refactor to consolidate ukvm asm parts
