+++ kernel/pvclock.c
@@ -77,38 +77,21 @@ struct pvclock_wall_clock {
 static volatile struct pvclock_vcpu_time_info pvclock_ti;
 static volatile struct pvclock_wall_clock pvclock_wc;
 
-static inline void
-x86_cpuid(uint32_t level, uint32_t *eax_out, uint32_t *ebx_out,
-        uint32_t *ecx_out, uint32_t *edx_out)
-{
-    uint32_t eax_, ebx_, ecx_, edx_;
-
-    __asm__(
-        "cpuid"
-        : "=a" (eax_), "=b" (ebx_), "=c" (ecx_), "=d" (edx_)
-        : "0" (level)
-    );
-    *eax_out = eax_;
-    *ebx_out = ebx_;
-    *ecx_out = ecx_;
-    *edx_out = edx_;
-}
-
 uint64_t pvclock_monotonic(void) {
     uint32_t version;
     uint64_t delta, time_now;
 
     do {
         version = pvclock_ti.version;
-        __asm__ ("mfence" ::: "memory");
+        cpu_mfence_memory();
         delta = cpu_rdtsc() - pvclock_ti.tsc_timestamp;
         if (pvclock_ti.tsc_shift < 0)
             delta >>= -pvclock_ti.tsc_shift;
         else
             delta <<= pvclock_ti.tsc_shift;
-        time_now = mul64_32(delta, pvclock_ti.tsc_to_system_mul) +
+        time_now = cpu_mul64_32(delta, pvclock_ti.tsc_to_system_mul) +
             pvclock_ti.system_time;
-        __asm__ ("mfence" ::: "memory");
+        cpu_mfence_memory();
     } while ((pvclock_ti.version & 1) || (pvclock_ti.version != version));
 
     return time_now;
@@ -124,10 +107,10 @@ static uint64_t pvclock_read_wall_clock(void)
 
 	do {
 		version = pvclock_wc.version;
-		__asm__ ("mfence" ::: "memory");
+		cpu_mfence_memory();
 		wc_boot = pvclock_wc.sec * NSEC_PER_SEC;
 		wc_boot += pvclock_wc.nsec;
-		__asm__ ("mfence" ::: "memory");
+		cpu_mfence_memory();
 	} while ((pvclock_wc.version & 1) || (pvclock_wc.version != version));
 
 	return wc_boot;
@@ -141,7 +124,7 @@ int pvclock_init(void) {
      * Prefer new-style MSRs, and bail entirely if neither is indicated as
      * available by CPUID.
      */
-    x86_cpuid(0x40000001, &eax, &ebx, &ecx, &edx);
+    cpu_x86_cpuid(0x40000001, &eax, &ebx, &ecx, &edx);
     if (eax & (1 << 3)) {
         msr_kvm_system_time = 0x4b564d01;
         msr_kvm_wall_clock = 0x4b564d00;
@@ -156,16 +139,13 @@ int pvclock_init(void) {
 
     printf("Solo5: Clock source: KVM paravirtualized clock\n");
 
-    __asm__ __volatile("wrmsr" ::
-        "c" (msr_kvm_system_time),
-        "a" ((uint32_t)((uintptr_t)&pvclock_ti | 0x1)),
-        "d" ((uint32_t)((uintptr_t)&pvclock_ti >> 32))
-    );
-    __asm__ __volatile("wrmsr" ::
-        "c" (msr_kvm_wall_clock),
-        "a" ((uint32_t)((uintptr_t)&pvclock_wc)),
-        "d" ((uint32_t)((uintptr_t)&pvclock_wc >> 32))
-    );
+    cpu_wrmsr(msr_kvm_system_time,
+              (uint32_t)((uintptr_t)&pvclock_ti | 0x1),
+              (uint32_t)((uintptr_t)&pvclock_ti >> 32));
+    cpu_wrmsr(msr_kvm_wall_clock,
+              (uint32_t)((uintptr_t)&pvclock_wc),
+              (uint32_t)((uintptr_t)&pvclock_wc >> 32));
+
     /* Initialise epoch offset using wall clock time */
     wc_epochoffset = pvclock_read_wall_clock();
