+++ kernel/ukvm/io.c
@@ -30,7 +30,7 @@ int solo5_net_write_sync(uint8_t *data, int n)
     wr.ret = 0;
 
     outl(UKVM_PORT_NETWRITE, ukvm_ptr(&wr));
-    cc_barrier();
+    cpu_cc_barrier();
 
     return wr.ret;
 }
@@ -44,7 +44,7 @@ int solo5_net_read_sync(uint8_t *data, int *n)
     rd.ret = 0;
 
     outl(UKVM_PORT_NETREAD, ukvm_ptr(&rd));
-    cc_barrier();
+    cpu_cc_barrier();
 
     *n = rd.len;
     return rd.ret;
@@ -56,7 +56,7 @@ char *solo5_net_mac_str(void)
     volatile struct ukvm_netinfo info;
 
     outl(UKVM_PORT_NETINFO, ukvm_ptr(&info));
-    cc_barrier();
+    cpu_cc_barrier();
 
     memcpy(mac_str, (void *)&info, 18);
     return mac_str;
@@ -73,7 +73,7 @@ int solo5_blk_write_sync(uint64_t sec, uint8_t *data, int n)
     wr.ret = 0;
 
     outl(UKVM_PORT_BLKWRITE, ukvm_ptr(&wr));
-    cc_barrier();
+    cpu_cc_barrier();
 
     return wr.ret;
 }
@@ -88,7 +88,7 @@ int solo5_blk_read_sync(uint64_t sec, uint8_t *data, int *n)
     rd.ret = 0;
 
     outl(UKVM_PORT_BLKREAD, ukvm_ptr(&rd));
-    cc_barrier();
+    cpu_cc_barrier();
 
     *n = rd.len;
     return rd.ret;
@@ -99,7 +99,7 @@ int solo5_blk_sector_size(void)
     volatile struct ukvm_blkinfo info;
 
     outl(UKVM_PORT_BLKINFO, ukvm_ptr(&info));
-    cc_barrier();
+    cpu_cc_barrier();
 
     return info.sector_size;
 }
@@ -109,7 +109,7 @@ uint64_t solo5_blk_sectors(void)
     volatile struct ukvm_blkinfo info;
 
     outl(UKVM_PORT_BLKINFO, ukvm_ptr(&info));
-    cc_barrier();
+    cpu_cc_barrier();
 
     return info.num_sectors;
 }
@@ -119,7 +119,7 @@ int solo5_blk_rw(void)
     volatile struct ukvm_blkinfo info;
 
     outl(UKVM_PORT_BLKINFO, ukvm_ptr(&info));
-    cc_barrier();
+    cpu_cc_barrier();
 
     return info.rw;
 }
