+++ kernel/virtio/cpu.S
@@ -21,15 +21,30 @@
 .text
 .code64
 
-ENTRY(cpu_halt)
-    cli
-    hlt
-    jmp cpu_halt
-END(cpu_halt)
-
-ENTRY(cpu_rdtsc)
-    rdtsc
-    shl $32, %rdx
-    add %rdx, %rax
+ENTRY(cpu_idt_load)
+    lidt 0(%rdi)
     ret
-END(cpu_rdtsc)
+END(cpu_idt_load)
+
+ENTRY(cpu_gdt_load)
+    lgdt 0(%rdi)
+    ret
+END(cpu_gdt_load)
+
+ENTRY(cpu_tss_load)
+    ltr %di
+    ret
+END(cpu_tss_load)
+
+ENTRY(cpu_sse_enable)
+    mov %cr0, %rax
+    and $0xfffb, %ax                    /* clear CR0.EM */
+    or $0x2, %ax                        /* set CR0.MP */
+    mov %rax, %cr0
+
+    mov %cr4, %rax
+    or $(3<<9), %ax                     /* set CR4.OSFXSR and CR4.OSXMMEXCPT */
+    mov %rax, %cr4
+
+    ret
+END(cpu_sse_enable)
