+++ kernel/virtio/cpu_vectors.S
